File view behaves like Windows Explorer
The file view displays the project file tree in the same manner as Windows
Explorer does. Tree items use native system icons and shell overlays as their
image which allows to track changes to folders and files when an SCM shell
extension like Tortoise* is used. Then, tree images are automatically updated
when a file or folder is modified or committed within or outside TeXnicCenter.
The shell context menu for each project file can accessed directly within the
view.

Missing files are emphasized using a red font color. When opening a project with
missing files it is ensured that the file items in the tree are not collapsed
and are visible to the user.

As a side note, shell extensions require the same bitness as the host process,
i.e. the application itself. For instance, a 32-bit version of TeXnicCenter will
not be able to display overlay icons of a 64-bit shell extension or vice versa.