Fixed crash in AutoCompleteDlg.
The reset() function was used to replace the same resource, meaning:
ptr(A).reset(A)
This led to a crash.
The assumption is that reset() releases A first,
and then assigns the now invalid resource to the unique pointer.

Since nobody calls CreateListBox elsewhere, it was an easy fix.
The creation of the dialog is only handled in CreateListBox now,
no resetting elsewhere.